Add new model concern Communitable with relation, callback and associate community.
